requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
drf-yasg>=1.21
django-cors-headers>=4.3
pandas>=2.0
python-dotenv>=1.0
Pillow>=10.0
